src/commands: Factor each command into a module
Consolidates all of the code which implements each command into a
separate module per command.

This should both make the codebase easier to understand and make it
simpler to add new commands since everything for a particular command is
now defined in the same command-specific module.
